+++ lib/screens/home_inspection_screen/components/room_detail_model_form.dart
@@ -9,7 +9,7 @@ import 'package:house_review/models/room_purpose.dart';
 typedef void RoomDetailFormSave(Room value);
 
 class RoomDetailModelForm extends StatefulWidget {
-  final int buildingStorey;
+  final double buildingStorey;
   final RoomDetailFormSave onFormSave;
   final Room toEditValue;
